refactor(meta/.../battleshipt): BREAKING CHANGE
BREAKING CHANGE

Made a module
